feat: Add recurring transactions card to home screen
This commit introduces a "Recurring Transactions" card to the home screen, which displays a list of all active recurring income and expenses. The underlying logic has been updated to materialize these recurring transactions for the current month's balance and expense calculations.

### Key Changes:

-   **Home Screen UI (`home_page.dart`):**
    -   Adds a new "Recurring Transactions" card that lists active recurring items for the month.
    -   Refactors UI cards into a reusable `HomeCard` widget for consistency.
    -   Improves the layout and styling of the balance, expenses, and "no expenses" sections.
    -   Adds a copyright footer.

-   **ViewModel & State (`home_viewmodel.dart`, `home_state.dart`):**
    -   Integrates `GetActiveRecurringTransactionsUseCase` to fetch and display recurring transactions.
    -   Implements logic (`isActiveForMonth`) to determine if a recurring transaction should be included in the current month's calculations.
    -   Materializes recurring transactions into regular `Transaction` objects to be included in the monthly balance and category summaries.
    -   Adds `recurring` transactions list to `HomeState`.

-   **Use Case & Domain (`get_active_recurring_transaction.dart`, `recurring_transaction_rules.dart`):**
    -   Introduces `GetActiveRecurringTransactionsUseCase` to fetch all active recurring transactions.
    -   Adds `recurring_transaction_rules.dart` to encapsulate the logic for checking if a transaction is active in a given month.

-   **Widgets (`recurring_tile.dart`, `home_card.dart`):**
    -   Creates `recurring_tile.dart` to display individual recurring transaction details.
    -   Creates a generic `HomeCard` widget to standardize the card UI across the home screen.

-   **Data & Database:**
    -   Removes the `big_decimal` dependency and refactors amount handling to use `int` (cents) across `Transaction`, `RecurringTransaction`, and related mappers for simplicity and consistency.
    -   Updates the `recurring_transactions` table schema, changing the `amount` column from `TEXT` to `INTEGER`.

-   **Refactoring & Fixes:**
    -   Refactors `transaction_label_resolver.dart` to separate emoji generation from label text.
    -   Adds `dispose()` method in `transaction_page.dart` to manage controllers.
    -   Adds `textAlign` property to the `CustomText` widget.
    -   Adds new internationalization strings for recurring transactions.

+++ lib/feat/transaction/domain/transaction.dart
@@ -1,10 +1,9 @@
-import 'package:big_decimal/big_decimal.dart';
 import 'package:finances_control/feat/transaction/domain/category.dart';
 import 'package:finances_control/feat/transaction/domain/enum_transaction.dart';
 
 class Transaction {
   final int? id;
-  BigDecimal amount;
+  int amount;
   TransactionType type;
   Category category;
   DateTime date;
